Refactor sync loadInstanceConfig
